+++ src/org/xjge/core/DGNoclip.java
@@ -9,7 +9,8 @@ import java.text.DecimalFormat;
  */
 class DGNoclip extends DebugGroup {
 
-    private final DecimalFormat pattern = new DecimalFormat("0.0");
+    private final DecimalFormat pattern1 = new DecimalFormat("0.0");
+    private final DecimalFormat pattern2 = new DecimalFormat("0");
     
     public DGNoclip(String title) {
         super(title);
@@ -23,16 +24,16 @@ class DGNoclip extends DebugGroup {
             
             switch(i) {
                 case 0 -> {
-                    output.get(i).append("POSITION:  (").append(pattern.format(noclip.position.x)).append(", ")
-                                                        .append(pattern.format(noclip.position.y)).append(", ")
-                                                        .append(pattern.format(noclip.position.z)).append(")");
+                    output.get(i).append("POSITION:  (").append(pattern1.format(noclip.position.x)).append(", ")
+                                                        .append(pattern1.format(noclip.position.y)).append(", ")
+                                                        .append(pattern1.format(noclip.position.z)).append(")");
                 }
                 case 1 -> {
-                    output.get(i).append("DIRECTION: (").append(pattern.format(noclip.direction.x)).append(", ")
-                                                        .append(pattern.format(noclip.direction.y)).append(", ")
-                                                        .append(pattern.format(noclip.direction.z)).append(")");
+                    output.get(i).append("DIRECTION: (").append(pattern1.format(noclip.direction.x)).append(", ")
+                                                        .append(pattern1.format(noclip.direction.y)).append(", ")
+                                                        .append(pattern1.format(noclip.direction.z)).append(")");
                 }
-                case 2 -> output.get(i).append("SPEED:     ").append("");
+                case 2 -> output.get(i).append("SPEED:     ").append(pattern2.format(noclip.speed * 100)).append("%");
             }
         }
     }