+++ src/org/xjge/core/XJGE.java
@@ -62,9 +62,6 @@ import org.xjge.graphics.PostProcessShader;
  */
 public final class XJGE {
     
-    static float noclipSpeedFactor = 1.0f;
-    
-    private static boolean noclipEnabled;
     private static boolean showLightSources;
     
     static GLProgram depthProgram;
@@ -737,21 +734,6 @@ public final class XJGE {
      */
     public static GLProgram getDefaultGLProgram() {
         return glPrograms.get("default");
-    }
-    
-    /**
-     * Sets the speed at which the camera will move while noclip mode is 
-     * enabled. This includes base speed and boost speed.
-     * <p>
-     * More specifically, the factor number provided here will influence the 
-     * base speed of the camera. That is, lower values will make the camera 
-     * move slower, whereas higher values will make it move faster. By default
-     * the cameras speed factor is 1.
-     * 
-     * @param factor the value that will effect the cameras movement speed
-     */
-    public static void setNoclipSpeedFactor(float factor) {
-        noclipSpeedFactor = clampValue(0, Float.MAX_VALUE, factor);
     }
     
 }